Sowmya| moves functions to drug chart utils
