+++ src/features/DisplayControls/DrugChart/styles/DrugListCell.scss
@@ -1,4 +1,4 @@
-.drug-name {
+.drug-chart-drug-name {
   max-width: 85%;
   white-space: nowrap;
   overflow: hidden;
@@ -6,6 +6,9 @@
   font-weight: 500;
   color: #393939;
   font-style: normal;
+}
+.drug-name-container {
+  display: flex;
 }
 .dosage {
   font-size: 12px;
@@ -17,4 +20,11 @@
 }
 .note-icon-container {
   margin-left: 5px;
+}
+.tooltip-content-separater {
+  height: 1px;
+  background-color: #bbb8b8;
+  width: 100%;
+  margin-top: 10px;
+  margin-bottom: 10px;
 }
